RESTRICTION: (using decorators) restricting unauthenticated users to create and delete rooms. Restricting unauthorized users from editing and deleting other user's rooms

+++ base/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 from .models import Room, Topic
 from .forms import RoomForm
@@ -57,6 +59,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -69,9 +72,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Invalid user')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -83,8 +90,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Invalid user')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
